Fix #testSplitscreenLandscapeAppOrientationRequests
In [1], we extracted RotationSession out of ActivityManagerTestBase,
and always create new WindowManagerStateHelper if we want to wait
for rotation finishing. However, previously the RotationSession
used the same WindowManagerStateHelper as the ActivityManagerTestBase
so AppConfigurationTests assume the WindowManagerStateHelper
was updated and led to verification failure on rotation.

This CL passes WindowManagerStateHelper to RotationSession so
ActivityManagerTestBase could use the existed WindowManagerStateHelper.

[1]: I0089edcaa71d7ec61f3a65929978b78f45af6810

Test: atest AppConfigurationTests
fixes: b/275547150
Change-Id: I2580c8c6342ee289eaa295f9b5aaa98aafddecb2