Run CtsContentTestCases with adb root disabled
CtsContentTestCases has tests that verify a runtime receiver
registered with the RECEIVER_NOT_EXPORTED flag cannot receive
broadcasts sent from external apps; the shell identity is used
as the identity that sends the broadcast since its UID (2000)
is not allowed to send broadcasts to unexported receivers.
However, on eng builds and others where adb is running as root,
the UID of the shell identity is 0, and this UID is allowed to
send broadcasts to unexported receivers. This commit uses the
RootTargetPreparer to ensure adb is not run as root during the test.

Fixes: 273644316
Test: atest CtsContentTestCases
Change-Id: I8f36bba38dfba5b5f16b4c7ed260b397933d2726