Test content:// URI in MediaPlayerTest CTS
This test uses FileProvider to create a content:// URI that points to a
local file, and passes this URI to MediaPlayer.setDataSource(Context,
Uri), then checks that playback completes successfully.

This URI is resolved by MediaPlayer.attemptDataSource before a binder
call into the mediaserver process.

MUST_SLEEP since this is the way all these othe tests are written

Test: atest MediaPlayerTest#testPlayContentUri
Bug: 274687614
Change-Id: I8435d9a6123d9266545f3e090c9bb6bc912124da