Wifi: increase the DPP test timeout values
This CL includes two changes,
1. Increase the timeout values for the tests which expect
an error from supplicant. This is because supplicant retries
the Tx frame 5 times and may set 10 seconds wait time for
scheduling the Tx frames.
2. After the test timeout, clean up the DPP resources in WifiStack
before checking the test results. This is to clean up the resources
in case of test failures(Before assert). Also added a 2 seconds
wait time for the cleanup call to reach the WiFi stack and execute.

Bug: 276188248
Test: atest EasyConnectStatusCallbackTest
Change-Id: If047fc2a24e8e1c1ef67a130cc6a1321c1c70540