telecom & telephony feature check in hostside test
The current hostside test for Telecom  stats are
failing on devices which have TELECOM feature but not TELEPHONY

Since the tests utilize existing CTS telecom tests that check for
both TELECOm and TELEPHONY feature, the hostside test fail because
the device side CTS test is skipped.

For now, Bring the hostside test feature check in sync with the
CTS tests that it is leveraging.

build/test on tablet https://android-build.googleplex.com/builds/abtd/run/L30500000959799138

Bug: 276396395
Test: Atest
Change-Id: I23e504e3ec57ddd28d765d712be8b7ad71aa23c8