[MockModem Data] Fix the wrong data call response for IMS
Symptom: Cuttlefish only has internet data call, but in our test case, we use the CHT carrier, and CHT will establish the internet and ims data calls. Since the ims is not supported by the original carrier, we report the initial value of ims data call, which will cause this symptom.

- Add a list to store the supported capability
- Add condition to check the capability if it is supported

Bug: 271764736
Test: Test on pixel6/7 Cuttle Fish passed
      atest TelephonyManagerTestOnMockModem
Change-Id: I318043dac8d7dc23c5b642fda9f6b4363155fbd9