Wait for broadcast idle after profile create
This change adds a wait for broadcast queues to be idle before starting
all the tests. This fixes test failures caused by adding account
through AccountManagerService before RegisteredServicesCache has
received the broadcast of android.intent.action.PACKAGE_ADDED. A timeout
of 2 min has been added to the "am wait-for-broadcast-idle" to ensure
the failure is time bound.

Bug: 276597375
Test: atest ContactsSharingTest

Change-Id: If9362fdf4b4788e91321ad9d325ec3008c1d6d51