Skip BT LE tests if feature is not supported
Skip BT LE test cases if FEATURE_BLUETOOTH_LE is not supported.
Currently for several testcases it is only checked if
FEATURE_BLUETOOTH is supported, but it is also possible that
a device only support BT classic and not BT LE (and vice versa).
Skip BT LE tests if feature is not supported.

Note: Unlike the CL merged in android12-tests-dev, this CL uses
Assume#assumeTrue in order to prevent code duplication.

Bug: 175314662
Test: Ran `atest <testClass1> <testClass2> ...`
Change-Id: I2cc8146800d48f4b07cf3741cc28d674774ca7c9
Signed-off-by: Timo Gienger <timo.gienger@elektrobit.com>