Add Test Cases Around Dynamic View Adds/Removals
Add additional cursor related tests around situations where EditText
views are dynamically added and removed from the layout.  This is in
response to a bug discovered by external dev aosp/2540031 where the
cursor does not resume blinking when an EditText view that had focus is
removed from the layout and then added back.

Bug: 278907680
Test: Manual testing.  Ran all tests in the EditTextTest class.
Change-Id: I4086ccb5a24e17bfe10d746ea088489978a3b655