WebView: deflake WebSettingsTest.testLocalImageLoads
This helper method would previously return too soon if the page title
was initialized to something other than "0". Now the helper method
will wait until the page title can be parsed as an integer > 0.

testLocalImageLoads() used a different HTML page which has a page title
that defaults to "TEST FAILED" which is why that test case was impacted.

Fixes: 277317623
Test: atest CtsWebkitTestCases:WebSettingsTest
Test: atest CtsWebkitTestCases:WebSettingsTest#testLocalImageLoads --rerun-until-failure 200
Change-Id: I5087e07341309972a5e701deabec3241993ade25