Move HorizontalScrollView to middle
Fixes: 279501424

The HorizontalScrollView that was used for testing the stretch
was up against the side of the screen. Even though the drag
was from the center of the HorizontalScrollView, the View was
very small and the touch event was being treated as a system
gesture.

The HorizontalScrollView was moved to the center of the
screen to avoid the edges.

Test: HorizontalScrollView on raven
Change-Id: Ic690da8fce97af4f0829abec370cec48a17bad6e