Update tests to support sticky broadcasts resent with UID
Whenever a receiver registers for a sticky broadcast, the sticky
broadcast was resent with a UID of -1 to allow the receiver to
immediately receive the broadcast. This worked prior to Android
T since all receivers were exported by default and could receive
broadcasts from any UID on the system, but with the introduction of
the RECEIVER_NOT_EXPORTED flag, sticky broadcasts from the local
app or system UID could not be delivered to a newly registered
receiver due to this UID of -1 not meeting the requirements for an
unexported runtime receiver. Android U resolves this by maintaining
and reusing the UID of the app that sent the sticky broadcast; this
commit updates the existing tests for sticky broadcasts to verify
this fix in Android U.

Bug: 244351792
Test: atest ContextTest
Change-Id: Ib5c6f7f9390fbcb90f8bb8ed9ac4e3cbf8b4f0e5