Fix testChangeFontScale flakiness
The test was flaky because there was a timing that the font scale
may not be dumped to the journal yet after the Activity#onCreate
is called. In that case, the test would fail because of not getting
the activity font scale.

Also waiting for #onResume callback to avoid the race.

Bug: 280026993
Test: atest ConfigChangeTests
Change-Id: Iff75449d6781f70d0d2bed22d3e2f3446c212e34