Modify MANAGE_ACCESSIBILITY to accept roles
This is the first step in allowing Exo and the Chromebook to act
on streamed screen content on behalf of its accessibility users.

It was assumed Exo would have access to this permission, but since
it's not signed by the platform, the signature-level permission is
not automatically granted. This fix allows only pre-installed,
privileged apps that have this role, and which have asked the user
for consent to stream and communicate with apps on the mobile device.

Bug: 280661435
Test: atest AccessibilityDisplayProxyTest, PermissionPolicyTest,
CtsCompanionDeviceManagerUiAutomationTestCases, CtsRoleTestCases

Change-Id: I3534a5bf9d8e57aff1a9b4fcb8c285ee3c13b4ae