Clarify steps in NotificationDismissVerifierActivity
The Notification title was mismatching from the test description,
because it was not set for the custom title was not set for the public layout of the notification.

Updated the title of the Notification, the title is always set and
always the same, because we don't need to distinguish if the public
layout is shown -> there is a diffent test for that.
Updated the wording of the test steps to avoid double negatives.

Fixes: 281452694
Test: manually execute NotificationDismissVerifierActivity
Change-Id: Id958cecf5379f819e0803010eecf72e8407b1de8