Adds not-smart Sleep after taking a screenshot.
On some devices taking a screenshot causes a popup dialog to
appear which can cause subsequent tests to fail if they run
immediately after this test. This dialog is not required to follow any
set of rules (and is outside Accessibility's control), so we cannot 
reliably wait for any well-defined condition like a window to appear
and disappear.

Instead, lets try to capture "most" cases by adding a simple
sleep after the action, and hope that most devices auto-hide the
within this limit.

Bug: 281697280
Change-Id: Iec958a713831597842365ddfa832a6d6692bd68f