Remove display windowing mode assumption
aosp/1413493 uses isInMultiWindowMode() to detect if the display
windowing mode is a vendor-defined windowing mode, which is not
appropriate because isInMultiWindowMode() returns true for freeform
mode as well. The method is not designed for detecting
WINDOWING_MODE_MULTI_WINDOW but for “multi-window mode” which includes
every mode except fullscreen.
In freeform mode, the activity lifecycle changes normally to RESUMED
after moving the task to top from bottom (behind home) so “app launch
count” should increment.
To ensure that the test is not affected by the display windowing mode,
this CL forces all activities to launch in fullscreen mode in this test
case.
Launching TEST_APP_PKG in fullscreen mode also allows us to remove the
pressHome() call because the top activity in mTargetPackage goes STOP by
TEST_APP_PKG. This is beneficial for devices without the home activity
(see config_noHomeScreen).

Test: atest testAppLaunchCount on aosp_cf_x86_64_pc and aosp_cf_x86_64_phone
Bug: 278966478
Change-Id: I6a4245e4798cbec0117d3cc84c0c4ff0a4130534