Opt out of foldable configuration runs
Currently, the test code get the text from textView.getLayout(). But
it is null from one of foldable mode. No clear clue to find why it
happens, we need more time to check. Because we don't have logic
difference between each foldable mode, it should not be a big problem
to opt out from foldable configuration.

Bug: 281961927
Fixes: 281950429
Test: atest CtsTranslationTestCases
Change-Id: I1ac0faf96372e90921b943c90789f9f469867088