[VoiceInteraction] Update HotwordDetectionServiceBasicTest
- Update testMultipleDetectors_onDetectFromDspAndMic_success,
testHotwordDetectionService_createDetectorTwiceQuickly_triggerSuccess
and testHotwordDetectionService_concurrentCapture
to verify can read audio and the data is not zero in
HotwordDetectionService and keep @RequiresDevice on it.
- Remove @RequiresDevice for some test cases
In emulator, if we have the annotation, it causes metric tests
fail with the error message "No tests were run on the device".
Keep the testMultipleDetectors_onDetectFromDspAndMic_success to
verify it. Remove it on the other test cases.
- Remove the no data check for the test cases without @RequiresDevice
It causes some test cases failed in emulator

Test: atest CtsVoiceInteractionTestCases on both emulator and
      the physical device
Bug: 281550925
Change-Id: I25c73e3ab7e94fae03bc6d039839a89c4e756e0d