Refactor neverallow tests to be parameterized
The neverallow tests were previously generated using Python which parsed
the policy when CTS was built and created a large amount of Java tests.
Modifications of the test code were cumbersome. Move to a parametrized
test which loads the platform policy when executing the test.

The template from SELinuxNeverallowTestFrame.py has been barely modified
beyond the necessary variable renaming.

The generating code from SELinuxNeverallowTestGen.py has been translated
to the generateRules() method. The same processing and regular
expressions are used. The same fail-safe guards are in place:
  - Test that each section ("BEGIN_xxx .. END_xxx") is open before being
    closed.
  - Test that each section is closed at the end of the processing.
  - (new) Test that at least one test was generated.

Further cleaning up is now possible (for instance moving the conditions
to an enum), but this is not included to facilitate the review.

Bug: 274037630
Test: run cts-dev --module CtsSecurityHostTestCases --test android.security.cts.SELinuxNeverallowRulesTest
Test: Manually confirm that the same number of tests are being generated in the previous and new implementation.
Change-Id: I91cac981787cb52b65bab32e8f07f91413fc84b3