Fix null object exception in UiAutomation during testing
During the case test, the system was busy and the UIAutomation connection timed out, resulting in mUiautomation being empty. Finally, it caused the case to call its interface and report a null pointer exception.

According to the official API documentation, when the Uiautomation connection times out, the caller should retry

Bug: 281210369
Test: assert UserRestrictionsTest
      assert CtsDevicePolicyManagerTestCases

Change-Id: I0f0c9e5ce49402939e01473285c4effc00cbf17b