Change CapturedActivity to wait for expected number of frames
The previous code ran CapturedActivity for a specified amount of time.
However, the time to run has been increasing since there are devices
that run very slowly and the tests require enough frames to be
considered valid. Instead, change the test to run for specified number
of frames and not time. This way, on fast devices, the test can complete
sooner, where on slow devices, they can take longer and still pass.

This will allow the tests to be added to presubmit since they no longer
are required to run for 60s and instead can be reduced to under 20s for
most devices.

Change-Id: Ib9a020eb98b52695894e2fdb93092077f9b7f1c8
Test: SurfaceSyncGroupContinuousTest
Test: SurfacePackageFlickerTest
Test: AttachedSurfaceControlSyncTest
Test: ASurfaceControlBackPressureTest
Test: SurfaceViewSurfaceValidatorTest
Test: SurfaceViewSyncTest
Bug: 282169297