Fix RestrictedPermissionsTest
When an IMMUTABLE PendingIntent was used to create an IntentSender
which was passed into Session.commit(IntentSender statusReceiver),
the PendingIntent cannot be modified and the caller couldn't receive
EXTRA_STATUS or any installation results.

The commit() will check if an IMMUTABLE IntentSender parameter is
passed to the system, the CTS should update it. We should also avoid
using implicit intent so also update the test.

Fix: 283375466
Fix: 240618202
Test: atest RestrictedPermissionsTest
Change-Id: Ic55f5faee4562d0064c2623c041b57b1fdaa8543