Adding target_sdk_version to AppClonongStorageStatsTest.
The target sdk version for this test will be 33 (T), as the orignal fix for this was included in tm-qpr-dev.
Failures were found on running this test on udc-mainline-prod with T build: https://android-build.corp.google.com/test_investigate/?invocationId=I69500010161890322&testResultId=TR37028672132174918
Also, preventing the test to be run only on Android version U and above.

Change-Id: Iec773e77ea1e2ba2b3f43df7589fec7418bc5468
Test: atest AppCloningStorageHostTest
Bug: 269792226