Disable deprecated ABI warning window
A new alert window was added to warn when 32-bit apps are
launched on 64-bit devices. However, that interrupts some UI
tests when they were launched in `module-abi: x86` or
`module-abi: armeabi-v7a` mode. To fix it, we need to add
the command to disable the warning dialog for these tests.

Bug: 284412613
Test: atest
Change-Id: I6b20267bb51b8528d35af2298c48e40d01642b45