Temporarily enable traces for testing
The CDD doesn't enforce traces to be enabled (`persist.traced.enable`)
by default in all device types. In fact, only handheld is required to
have do to so.

Originally, the test assumes that traces is enabled by default, which
may not be true for all device types.

Improve that by temporarily enabling traces if it's not enabled, run
the tests and disable traces again if enabled by the test case.

MUST_SLEEP to wait for traces to complete

Bug: 285215697
Change-Id: I2f116debf9e55968c37c79b6e0576dba41be1322