Fix CTS flakes
The tests were flaky because we were storing the instance of the
TileService in TileService.onTileAdded and querying that instance for
different states. However, after the tile is added, it could be
bound/unbound multiple times, which made it necessary for state to be
stored outside of the instance of TileService (statically). However, as
this is not easy to work around with inheritance (and there are no
meta-classes) we do the following:

* Create a class with just static methods that tracks the current state
  of the test. This is effectively a Singleton.
* Reset this class at the beginning and end of the test.
* Keep track of the instance of the TileService only while it's
  listening (as this is guaranteed to be in the same bind). We cannot
  use onCreate/onDestroy, as they are not part of the TileService
  lifecycle.
* When the tile is added, keep track of which class was added (to verify
  it's the correct one).

MUST_SLEEP because there are animations

Fixes: 279975268
Fixes: 282667251
Test: atest CtsTileServiceTestCases

Change-Id: I32e31598e0430de0b279989a7aa585798bec32b9