Remove unnecessary cts tag.
Some modules defined in the build file are only data dependencies of cts test modules. In this case, they will be packaged to corresponding test modules. For example, if module A is a data dependency of test module B, the module A can be found under android-cts/testcases/B/ directory. Therefore it is unnecessary to tag data dependencies which will cause a duplicate folder, e.g. android-cts/testcases/A.

Change-Id: I1c9d9bec5d48e04c054c4522edd14514db1ceae4
Test: m cts
Bug: 283199104