Adds cts test on NotificationRankingUpdate
Adds a cts test to ensure NotificationRankingUpdates are delivered to
NotificationListeners without Binder Errors in the case that many
notifications are posted or removed in a short time.

To do this, we bind a TestNotificationListener, then
create ten apps and post 50 notifications from each
of them, then unbind all ten apps at the same time, counting calls
to the onPosted and onRemoved methods of the TestNotificationListener
with a CountDownLatch. We expect a minimum number of these calls
to be made.

We set persist.sysui.notification.ranking_update_ashmem with
adb shell setprop persist.sysui.notification.ranking_update_ashmem true
in AndroidTest.xml, as this is required for the test to pass.

Bug: 249848655
Test: atest tests/tests/notification/src/android/app/notification/current/cts/NotificationManagerTest.java#testRankingUpdateSentWithPressure
Change-Id: I35c1c53037e0265d86879062fa879bcdcdf3167d