Adds cts test on NotificationRankingUpdate
Adds a cts test to ensure NotificationRankingUpdates are delivered to
NotificationListeners without Binder Errors in the case that many
notifications are posted or removed in a short time.

To do this, we bind a TestNotificationListener, then
create eight apps and post 50 notifications from each
of them, then unbind all apps at the same time, counting calls
to the onPosted and onRemoved methods of the TestNotificationListener
with a CountDownLatch. We expect a minimum number of these calls
to be made.

We set persist.sysui.notification.ranking_update_ashmem with
adb shell setprop persist.sysui.notification.ranking_update_ashmem true
in AndroidTest.xml, as this is required for the test to pass.

In the previous submission of this test, we didn't clear the
notifications before unregistering the packages, which sometimes caused
the test to fail. Here, we clear all notifications first.
Also, we add sleeps between each service connection bind attempt, to
avoid low memory devices from experiencing superlinear bind times when
the service connections are all established at once.

Bug: 249848655
Test: atest tests/tests/notification/src/android/app/notification/current/cts/NotificationManagerTest.java#testRankingUpdateSentWithPressure
Change-Id: I79c6867fface5f356514cfcc0fd4827bc95a3db5