webkit: Update testCookieManagerBlockingUiThread to use http
This test was previously using HTTPS but started flaking out. It reported
that it used HTTPS for instant apps but we only execute this from testCookieManager
which only runs in AppModeFull.

We can't open a TCP socket in instant app mode so this reason is not required.

Mitigating this flaky test by switching it to HTTP.

Test: atest CtsHostsideWebViewTests
Bug: 286046288
Change-Id: If2bc8316030ae2c1228e670281b4febb4c44ca70