Fix the failing CTS ViewGroupTest after the hover icon changes
The previous CL ag/23109421 changed the return value of
ViewGroup#onResolvePointerIcon to null when the pointer is hovering
on the scroll bar.(Previously it returns POINTER_ARROW.) It untilized
the fallback behavior of pointer icon when the returned value is null.
For mouse, it will still display POINTER_ARROW. For stylus, it won't
show hover icon.
This CL updates the CTS tests to match the new logic.

Bug: 284223647
Test: atest ViewGroupTest
Change-Id: Ibf00b1f27fc35e2db9b23a49a4ed77b221b19e1f