Handle low targetSDK warning dialog right after installing
When we install a test app of an appropriate sdk, we can open it right
away to dismiss the low target SDK dialog, and not have to deal with it
during the actual test. Previously, we'd been doing a very quick check
every time we performed certain methods. This was flaky, and sometimes
missed the dialog. By checking for the dialog only at times we know it
will be shown, we can wait longer, to ensure it gets cleared

Fixes: 287968066
Test: atest CtsPermissionTestCases
Change-Id: Iab9917ae2d87310032d60ece20e25417a35a1bf5