[DO NOT MERGE] Effectively ignore testAppSearchSchema_toString
The string comparison is too restrictive and older tests will break if
we change the toString method at all. The testGetSchema* tests in
AppSearchSessionCtsTest are sufficient to ensure that every field set in
the schema builder makes it to the resulting schema.

Just do a loose check to make the test pass on T. Now we just make sure
all the previous expected lines are present in the new output.

The change is taken from aosp/2644598.

Test: testAppSearchSchema_toString on U devices
Bug: 287720061
Change-Id: I862977cced8b8d76ea481814596afa702e9f494b