Use Messenger to replace broadcast for interaction of test app
The broadcast latency may be non deterministic depending on how
ActivityManager schedules or whether the dispatching thread is busy.

To avoid occasional test failures by the latency, replace the 2-way
broadcast communication to Messenger that can reach the target
process directly:
 1. Test case puts its Messenger (C) to provider.
 2. Test app gets (C) from provider and puts its Messenger
    to provider (H) and tell (C) the app is ready.
 3. Test case gets (H) from provider then the test can
    send other command to control the test app with
    response data.

Here uses ContentProvider to establish the connection because
the protocol needs to support launching activity from shell
command, which is unable to send a Binder directly.

Bug: 288301550
Test: CtsWindowManagerDeviceTestCases
Change-Id: I9b930e45f484fcd133bb12cfb70033b5271f8d05