Fix testCropWithChildBoundingInsets test
There's a possible race where the bounds layer is first created and set
with no bounds in its own Transaction. When the draw pass occurs, the
bounds layer is updated to use the set bounds. However, this transaction
is applied using the buffer apply token so there's no order guarantee
with the first transaction. That means the first transaction that
contains no bounds could be applied after the transaction that contains
the real bounds. This would cause the test to fail since the crop would
end up getting removed.

Test: testCropWithChildBoundingInsets 100 times
Bug: 288339794
Change-Id: I7c6d8b8106b4e617db436a2b8d0cd855504d3dfd