HDMI: Fix failing HdmiCecPhysicalAddressTest CTS
In the testing environment the playback devices are connected to an HDMI Stub
that assigns the device a different PA than 1.0.0.0. Update the test
s.t. it stores the PA before the reboot.

Remove the unused variable `dutPhysicalAddress` and update the
documentation accordingly.

Using the `cec-phy-addr` option when running CTS will have no effect.

Bug: 290184848
Test: run cts -m CtsHdmiCecHostTestCases -t android.hdmicec.cts.common.HdmiCecPhysicalAddressTest#cect_10_1_2_1_RebootPhysicalAddress
Change-Id: I63e335b1aaf3e3973355f71b06c42810c928c0b8