Increase long key press to reduce flaky test
Bug: 288339642

TooltipTest.testLongKeyPressTooltipOnClickableView() uses
a long press on the ENTER key to trigger tooltip. The
timeout was set to the long press timeout. Since it
isn't possible to reproduce this with forrest, I'm
guessing that by sending the KEY_UP event exactly
after the timeout may cause the KEY_UP event to be
processed prior to the timeout of the long press.

This change doubles the long press time to give the
device a chance to process the long press, even on
a slower/busy device.

Test: ran test locally and on forrest

Change-Id: Ia73055f83cd9f6d616c33fb0b866f270f936d6a2