Allow some wiggle room in launched activity bounds in A11y CTS.
Some tests recently started failing with +/- 1 pixel differences.
I'm not sure the specific cause, but this data comes from various
sources (WindowManager vs. SurfaceFlinger) and the activitiy is
floating mid-screen so it may be rounding issues.

Also updates the log messages to log even when the window isn't
yet focused or active. There's no reason to exclude these messages,
and they're helpful in debugging tests that fail to make the
window active.

Bug: 291610218
Test: atest AccessibilityEmbeddedHierarchyTest
Test: use go/abtd to run many iterations of the tests
Change-Id: I5a86d4dac913564b90cb00779f993b7c3325af7d