Run BugreportManagerTest on HSUM
Removes the headless system user assumption from the tests,
and accounts for the user ID when parsing the bugreport file
names. Adds logic for handling remote bugreports on secondary
users by creating a system context for receiving the finished
broadcast, as remote bugreport requests will be sent to the
system user rather than the current user.

Test: adb shell cmd user set-system-user-mode-emulation headless,
      atest BugreportManagerTest
Bug: 267453822
Change-Id: Id719752cb4b430f72fb0a785cb966a916e233874