[RESTRICT MERGE] Modify assist testcase-Add NULL check
It is observed that this cts testcase(android.assist.cts.TextViewTest#testTextView)
fails in automotive usecases where multiple displays are used. If the displays are
virtual displays marked with FLAG_SECURE, as per AOSP documentation, the
assist structure returned is NULL, thereby causing the testcase to fail. This patch
fixes this bug.

Already, there is a check added for automotive cases, where it returns if a different
activity is received instead of what's expecting. This patch adds a NULL check for
assist structure in onHandleAssist(...) function, for handling the case of virtual
secure displays, so that the function returns without causing the testcase to fail.
Bug: 290968943

Change-Id: I4dd37dcd8485520f0a76b6f1f36256ab7325f2ad