Remove getBoundsToCheck from SurfacePackageFlickerTest
SurfacePackageFlickerTest wants to verify all the pixels in the window.
In that case, it can just use the default getBoundsToCheck which returns
null and will check the full bounds of the window.

This fixes a crash where the view framework bounds was slightly larger
than the window bounds, causing the test to crash when trying to check
out of bounds pixels in the VD.

Test: SurfacePackageFlickerTest
Bug: 292297177
Change-Id: I6b90f9bce5942b9eb44d94878c616eb63eb4ff50