Copy over ATS CarPowerManager tests to CTS
The tests in CarPowerManagerTest in the ATS don't really belong in ATS.
They should be a part of CTS since they are testing API. To begin the
migration, the test cases that aren't currently covered by the CTS
version of CarPowerManagerTest are copied over.

The testApplyNewPowerPolicy test case is adjusted to actually call the
applyPowerPolicy function from CarPowerManager instead of using a shell
command to call directly into CarPowerManagementService.

The CTS CarPowerManagerTest is also adjusted to enable/disable required
permissions at the beginning/end of the test instead of using the
execute shell command with permissions.

Missing @ApiTest annotations are added as well.

Bug: b/233786318
Test: atest CtsCarTestCases:CarPowerManagerTest
Change-Id: Ib9b8e00883419af570d307792175c3c7d5d08580