Check autobrightness enabled in BrightnessTest
This cl checks if a default brightnessconfiguration exists. This will
return null if automatic brightness controller is null, which only
happens if the oem has disabled autobrightness, not just the user.
We need to give this method permission to control display brightness
before calling this method.

If abc (and therefore brightnessconfiguration) doesnt exist, the
assumption fails and the test doesnt run.

Bug: 292023923
Test: atest BrightnessTest
Change-Id: I358678d731a95510fba7d4dd9d8dc29be3d2ff5c