DO NOT MERGE
WM : Add a visibility test in waitForEnterPip.

Under some conditions (high CPU Usage, launcher playing video content in
the background, etc.), the PIP takes a few seconds to display.
In this case, the testLaunchTaskByAffinityMatchSingleTask() might fail
to bring the PIP to full screen.
This is because waitForEnterPip returns too quickly.

This fix adds a condition in waitForEnterPip to avoid this behavior.
In this way, we wait for the associated task to be visible.

Bug: 295047609
Change-Id: I6063a3aaf665785717898d0115c773051fce8c0c
Merged-In: I6063a3aaf665785717898d0115c773051fce8c0c
Signed-off-by: Moncef LIMAM <moncef.limam-ext@sagemcom.com>
(cherry picked from commit 10cc03a40593109f1376d0513eda816c0b2ae844)